changing 960x600 to 1280x960 (720p) in index.html
this fixes the weird issue I had when exporting my 720p view using this template. I'd bet almost all users will be on desktops with 1080p or better monitors, so we can at least set the game window to 720p and still have room for the buttons underneath the game window.